# Search: Optimize dashboard permissions query