--- a/pkg/services/sqlstore/permissions/dashboard.go
+++ b/pkg/services/sqlstore/permissions/dashboard.go
@@ -26,52 +26,24 @@ func (d DashboardPermissionFilter) Where() (string, []interface{}) {
 	}
 
 	falseStr := d.Dialect.BooleanStr(false)
-
-	sql := `(
-		dashboard.id IN (
-			SELECT distinct DashboardId from (
-				SELECT d.id AS DashboardId
-					FROM dashboard AS d
-					LEFT JOIN dashboard AS folder on folder.id = d.folder_id
-					LEFT JOIN dashboard_acl AS da ON
-						da.dashboard_id = d.id OR
-						da.dashboard_id = d.folder_id
-					LEFT JOIN team_member as ugm on ugm.team_id = da.team_id
-					WHERE
-						d.org_id = ? AND
-						da.permission >= ? AND
-						(
-							da.user_id = ? OR
-							ugm.user_id = ? OR
-							da.role IN (?` + strings.Repeat(",?", len(okRoles)-1) + `)
-						)
-				UNION
-				SELECT d.id AS DashboardId
-					FROM dashboard AS d
-					LEFT JOIN dashboard AS folder on folder.id = d.folder_id
-					LEFT JOIN dashboard_acl AS da ON
-						(
-							-- include default permissions -->
-							da.org_id = -1 AND (
-							  (folder.id IS NOT NULL AND folder.has_acl = ` + falseStr + `) OR
-							  (folder.id IS NULL AND d.has_acl = ` + falseStr + `)
-							)
-						)
-					WHERE
-						d.org_id = ? AND
-						da.permission >= ? AND
-						(
-							da.user_id = ? OR
-							da.role IN (?` + strings.Repeat(",?", len(okRoles)-1) + `)
-						)
-			) AS a
+	sql := `dashboard.id IN (
+    	SELECT d.id FROM dashboard as d
+    	WHERE d.org_id IN (-1, ?)
+        AND (
+			d.has_acl = ` + falseStr + `
+        	OR (
+            	SELECT acl.dashboard_id FROM dashboard_acl AS acl WHERE d.id = acl.dashboard_id
+            	AND acl.permission >= ?
+            	AND (
+                	acl.user_id = ?
+                	OR acl.team_id IN (SELECT team_id FROM team_member WHERE team_member.user_id = ?)
+                	OR acl.role IN (?` + strings.Repeat(",?", len(okRoles)-1) + `)
+            	)
+        	)
 		)
-	)
-	`
+	)`
 
 	params := []interface{}{d.OrgId, d.PermissionLevel, d.UserId, d.UserId}
-	params = append(params, okRoles...)
-	params = append(params, d.OrgId, d.PermissionLevel, d.UserId)
 	params = append(params, okRoles...)
 	return sql, params
 }