**What this PR does / why we need it**:
When we are loading the dashboards, for a non-admin user, we need to check if the user has permissions on every dashboard before return anything.

The current query that we are doing to check that it takes more and more time every time that we create a dashboard (or folder) or add permissions of any kind.

The idea is to use Subqueries due for MySQL databases for example, JOINs load on memory the whole table before perform any search on it. So, every time that we add new information, this query starts to take more and more time to retrieve the information.

We have customers with more than 10k dashboards and more than 1k users and Grafana is really really slow loading the information.

The query that this PR improves is `/api/search` one.

**Tests**
It was tested with 13k dashboards + 2k users + 300 teams + ~130k of dashboard/folder permissions using Mac in localhost.

* SQLite3:
  * Dashboard page:
    * Now: 18.99 seconds
    * Before: 3.3 minutes
  * Homepage:
    * Now: 19 seconds
    * Before: 3.4 minutes

* MySQL:
  * Dashboard page:
    * Now: 18 seconds
    * Before: 3.2 minutes
  * Homepage:
    * Now: 19.6 seconds
    * Before: 3.1 minutes

* Postgres:
  * Dashboard page:
    * Now: 68 milliseconds
    * Before: 3.12 seconds
  * Homepage:
    * Now: 46 milliseconds
    * Before: 2.8 seconds

* I configured some folders and dashboards that can be access by one few users, roles or teams and test with different users.

**Other notes**
When you open the homepage or dashboard page maybe you don't feel that the pages are "faster" because it isn't the only query that the pages are doing. It's something that we need to investigate to improve it too.

**Notes for reviewers**
* It should be tested with an non-admin user because an admin has the whole access to dashboards and this query is skipped.
* If you can test it with a similar environment could be fine to check that I didn't miss something. The most important thing is that the behaviour is the same as before.
* I couldn't test it with MSSQL.

And really thanks to @undef1nd for her help to make the query optimised and readable as possible.